:recycle: Refactor BoosterPack in its own file
