+++ source/YugiAcademy/YuGiOh/views.py
@@ -1,5 +1,9 @@
 from django.shortcuts import render, redirect
-from .models import MonsterCard, SpellCard, TrapCard, BoosterPack, BoosterPackCard
+from cards.monster_card import MonsterCard
+from cards.spell_card import SpellCard
+from cards.trap_card import TrapCard
+from booster_packs.booster_pack import BoosterPack
+from booster_packs.booster_pack_card import BoosterPackCard
 from .forms import SpellCardForm, SearchBoosterPackForm, BoosterPackForm, BoosterPackCardForm
